+++ CHANGELOG.md
@@ -91,6 +91,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 - **hardis:work:save** and **hardis:project:deploy:sources:dx**: Improve runtime performances thanks to internalization of sfdx-essentials commands
 - **hardis:work:new**: Allow to add labels in property `availableTargetBranches`, using a comma. For examples, `- integration,Choose this branch if you are on the BUILD side of the project !`
 - **hardis:org:files:export**: Improve display with spinner
+- **hardis:project:generate:gitdelta**: Add option to generate package.xml related to a single commit
 
 ### Documentation